Extracted common expiration utility service
