+++ broker/src/main/java/io/moquette/broker/scheduler/Expirable.java
@@ -1,0 +1,7 @@
+package io.moquette.broker.scheduler;
+
+import java.time.Instant;
+
+public interface Expirable {
+    Instant expireAt();
+}
